requirements.txt
PyJWT==2.6.0
fastapi==0.88.0
uvicorn==0.20.0
SQLAlchemy==1.4.45
PyMySQL==1.0.2
pydantic==1.10.2
Jinja2==3.1.2
decouple==0.0.7
python-multipart==0.0.5
cryptography==38.0.4